pyproject.toml
[project]
name = "gemini-chatbot"
version = "0.1.0"
description = "A Streamlit chatbot powered by Google Gemini AI"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.30.0",
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]
